[TuneTest] * fix float/int issue
13:33:18.2296   File "/usr/lib/enigma2/python/Components/TuneTest.py", line 14, in tune
13:33:18.2311 TypeError: in method 'eDVBFrontendParametersSatellite_frequency_set', argument 2 of type 'int'
(cherry picked from commit ad36e768b3c80f38c7287e9aee4b2ad284c2a11c)

+++ lib/python/Components/TuneTest.py
@@ -11,8 +11,8 @@ class Tuner:
 		if self.frontend:
 			print("[TuneTest] tuning to transponder with data %s" % str(transponder))
 			parm = eDVBFrontendParametersSatellite()
-			parm.frequency = transponder[0] * 1000
-			parm.symbol_rate = transponder[1] * 1000
+			parm.frequency = int(transponder[0] * 1000)
+			parm.symbol_rate = int(transponder[1] * 1000)
 			parm.polarisation = transponder[2]
 			parm.fec = transponder[3]
 			parm.inversion = transponder[4]
